[C-8] Small closure refactor.

+++ internal/features/daily/service/correction_service.go
@@ -5,18 +5,24 @@ import (
 
 	"github.com/google/uuid"
 
+	"github.com/ivan-ca97/life/pkg/dayclosure"
+
 	"github.com/ivan-ca97/life/internal/features/daily/domain"
 	"github.com/ivan-ca97/life/internal/features/daily/ports"
 )
 
 type correctionService struct {
-	repository ports.CorrectionRepository
+	repository     ports.CorrectionRepository
+	closureChecker dayclosure.DayClosureChecker
 }
 
 var _ ports.CorrectionService = (*correctionService)(nil)
 
-func NewCorrectionService(repository ports.CorrectionRepository) *correctionService {
-	return &correctionService{repository: repository}
+func NewCorrectionService(repository ports.CorrectionRepository, closureChecker dayclosure.DayClosureChecker) *correctionService {
+	return &correctionService{
+		repository:     repository,
+		closureChecker: closureChecker,
+	}
 }
 
 func (s *correctionService) GetCorrection(userId uuid.UUID, date time.Time) (*domain.Correction, error) {
@@ -24,9 +30,25 @@ func (s *correctionService) GetCorrection(userId uuid.UUID, date time.Time) (*do
 }
 
 func (s *correctionService) UpsertCorrection(userId uuid.UUID, correction *domain.Correction) error {
+	closed, err := s.closureChecker.IsClosed(userId, correction.Date)
+	if err != nil {
+		return err
+	}
+	if closed {
+		return dayclosure.ErrDayClosed
+	}
+
 	return s.repository.UpsertCorrection(userId, correction)
 }
 
 func (s *correctionService) DeleteCorrection(userId uuid.UUID, date time.Time) error {
+	closed, err := s.closureChecker.IsClosed(userId, date)
+	if err != nil {
+		return err
+	}
+	if closed {
+		return dayclosure.ErrDayClosed
+	}
+
 	return s.repository.DeleteCorrection(userId, date)
 }
